spcript for scrapper fix

See merge request playmaker1/webapp!204

+++ connector/scripts/import_scrapper_objects.py
@@ -3,7 +3,7 @@ from django.contrib.postgres.search import SearchVector
 import django.core.exceptions
 from collections import Counter
 
-from .utils import unify_name, NEW_LNP_SOURCE, NEW_ADDITIONAL_LNP_SOURCE, create_mapper, get_mapper, MapperEntity
+from .utils import LNP_SOURCE, create_mapper, get_mapper, unify_team_name, unify_club_name
 from .base import BaseCommand
 from .restore_database import Command as RestoreDatabase
 from clubs.models import (
@@ -29,71 +29,7 @@ from connector.entities import (
     TeamEntity,
     BaseClubEntity,
 )
-
-PARENT_UUID_REQUIRED = ["Ekstraklasa", "Ekstraliga K", "1 Liga", "2 liga"]
-
-JUNIOR_LEAGUES = [
-    "A1",
-    "A2",
-    "B1",
-    "B2",
-    "C1",
-    "C2",
-    "D1",
-    "D2",
-    "E1",
-    "E2",
-    "F1",
-    "F2",
-    "G1",
-    "G2",
-    "CLJ U-19",
-    "CLJ U-18",
-    "CLJ U-17",
-    "CLJ U-15",
-]
-
-LEAGUE_HIGHEST_PARENT_NAME_MAPPER = {
-    "Ekstraklasa": "Ekstraklasa",
-    "Ekstraliga kobiet": "Ekstraliga K",
-    "Pierwsza liga": "1 Liga",
-    "Pierwsza liga kobiet": "1 Liga K",
-    "Druga liga": "2 liga",
-    "Druga liga kobiet": "2 Liga K",
-    "Trzecia liga": "3 Liga",
-    "Trzecia liga kobiet": "3 Liga K",
-    "Czwarta liga": "4 Liga",
-    "Czwarta liga kobiet": "4 Liga K",
-    "Piąta liga": "5 Liga",
-    "Klasa okręgowa": "Klasa Okręgowa",
-    "Klasa A": "A Klasa",
-    "Klasa B": "B Klasa",
-    "Klasa C": "C Klasa",
-    "CLJ U-19": "Clj U-19",
-    "CLJ U-18": "Clj U-18",
-    "CLJ U-17": "Clj U-17",
-    "CLJ U-15": "Clj U-15",
-    "Centralna Liga Juniorek U-17": "Clj U-17 K",
-    "Centralna Liga Juniorek U-15": "Clj U-15 K",
-    "A1": "Junior A1",
-    "A2": "Junior A2",
-    "B1": "Junior Młodszy B1",
-    "B2": "Junior Młodszy B2",
-    "C1": "Trampkarz C1",
-    "C2": "Trampkarz C2",
-    # "D1": "Młodzik D1 U-13", # We don't need them yet
-    # "D2": "Młodzik D2 U-12",
-    # "E1": "Orlik E1 U-11",
-    # "E2": "Orlik E2 U-10",
-    # "F1": "Żak F1 U-9",
-    # "F2": "Żak F2 U-8",
-    # "G1": "Skrzat G1 U-7",
-    # "G2": "Skrzat G2 U-6",
-    "Futsal Ekstraklasa": "Futsal Ekstraklasa",
-    "I Liga PLF": "I Liga PLF",
-    "II Liga PLF": "II Liga PLF",
-    "III Liga PLF": "III Liga PLF",
-}
+from mapper.enums import LEAGUE_HIGHEST_PARENT_NAME_MAPPER, JUNIOR_LNP_LEAGUES, PARENT_UUID_REQUIRED
 
 
 class Command(BaseCommand):
@@ -218,21 +154,12 @@ class Command(BaseCommand):
             result = base.filter(search_name=name)
             if l_highest_parent:
                 result = result.filter(league__highest_parent__name=l_highest_parent)
-            else:
-                try:
-                    direct = model.objects.get(name=name)
-                    return direct
-                except (
-                        django.core.exceptions.MultipleObjectsReturned,
-                        django.core.exceptions.ObjectDoesNotExist,
-                ):
-                    pass
-            if len(result) > 1:
+            if len(result) == 1:
+                return result[0]
+            elif len(result) > 1:
                 res_list = [res.name for res in result]
                 if res_list and res_list.count(res_list[0]) == len(res_list):
                     return result[0]
-            elif len(result) == 1:
-                model_obj = result[0]
             return model_obj
 
         try:
@@ -242,6 +169,8 @@ class Command(BaseCommand):
                     return model.objects.get(mapper=target_mapper)
                 elif model is Team:
                     return model.objects.get(mapper=target_mapper)
+            else:
+                return partial_search()
         except django.core.exceptions.ObjectDoesNotExist:
             return partial_search()
 
@@ -382,7 +311,7 @@ class Command(BaseCommand):
                         self.swap_lh_round("RW", target_league, local_season)
 
                     try:
-                        league_history = LeagueHistory.objects.get(
+                        LeagueHistory.objects.get(
                             mapper=get_mapper(play.id)
                         )
                     except django.core.exceptions.ObjectDoesNotExist:
@@ -391,37 +320,35 @@ class Command(BaseCommand):
                                 league=target_league,
                                 season=local_season,
                             )
+                            target_mapper_entity = league_history.mapper.get_entity(source=LNP_SOURCE,
+                                                                                    related_type="play")
+                            if (
+                                    target_mapper_entity
+                                    and target_mapper_entity.mapper_id != play.id
+                            ):
+                                additional_phrases = re.findall('"([^"]*)"', play.name)
+                                if additional_phrases:
+                                    new_child = additional_phrases[0].capitalize()
+                                    if new_child == "Ekstraliga":
+                                        new_child += f" {league_obj.name}"
+                                else:
+                                    new_child = "II"
+                                target_league, _ = League.objects.get_or_create(
+                                    name=new_child,
+                                    highest_parent=highest_parent_object,
+                                    parent=target_league,
+                                    scrapper_autocreated=True,)
                         except django.core.exceptions.ObjectDoesNotExist:
-                            league_history = LeagueHistory.objects.create(
-                                league=target_league,
-                                season=local_season,
-                                mapper=create_mapper(
-                                    NEW_LNP={"id": play.id, "desc": "LNP play uuid"},
-                                    NEW_ADDITIONAL_LNP={"id": league_obj.id, "desc": "LNP league uuid (highest parent)"}
-                                ),
-                                league_name_raw=play.name,
-                            )
-                    target_mapper_entity = league_history.mapper.get_entity(source=NEW_LNP_SOURCE)
-                    if (
-                        league_history.mapper
-                        and target_mapper_entity
-                        and target_mapper_entity.mapper_id != play.id
-                    ):
-                        lh_mapper_entity = league_history.mapper.get_entity(source=NEW_LNP_SOURCE)
-                        curr_teams_count = len(
-                            self.service.get_play_teams(lh_mapper_entity.mapper_id)
-                        )
-                        new_teams_count = len(self.service.get_play_teams(play.id))
-                        if new_teams_count < curr_teams_count:
-                            continue
-                        lh_mapper_entity.mapper_id = play.id
-                        lh_mapper_entity.save()
-                        if highest_parent in PARENT_UUID_REQUIRED:
-                            lh_mapper_additional_entity = league_history.mapper.get_entity(source=NEW_ADDITIONAL_LNP_SOURCE)
-                            lh_mapper_additional_entity.mapper_id = league_obj.id
-                            lh_mapper_additional_entity.save()
-                        league_history.league_name_raw = play.name
-                        league_history.save()
+                            pass
+                    LeagueHistory.objects.create(
+                        league=target_league,
+                        season=local_season,
+                        mapper=create_mapper(
+                            {"id": play.id, "related_type": "play", "database_source": "scrapper_mongodb", "desc": "LNP play uuid"},
+                            {"id": league_obj.id, "related_type": "league", "database_source": "scrapper_mongodb", "desc": "LNP league uuid (highest parent)"}
+                        ),
+                        league_name_raw=play.name,
+                    )
 
     def map_teamhistory(self, team_histories: List[TeamHistoryEntity]) -> None:
         """
@@ -440,9 +367,9 @@ class Command(BaseCommand):
                     ]
                 except KeyError:
                     continue
-                team_name = unify_name(team.name, False)
+                team_name = unify_team_name(team.name)
                 seniority = (
-                    self.JUNIOR if team_league.name in JUNIOR_LEAGUES else self.SENIOR
+                    self.JUNIOR if team_league.name in JUNIOR_LNP_LEAGUES else self.SENIOR
                 )
                 gender = (
                     self.FEMALE
@@ -456,7 +383,7 @@ class Command(BaseCommand):
                     unique_id=team_history.obj_id,
                 )
 
-                club_name = unify_name(club.name)
+                club_name = unify_club_name(club.name)
                 club_details = self.service.get_club_details(club.id)
                 club_address = club_details.address
                 voivo = club_details.voivodeship
@@ -472,7 +399,7 @@ class Command(BaseCommand):
                 def create_club():
                     return Club.objects.create(
                         name=club_name,
-                        mapper=create_mapper(NEW_LNP={"id": club.id, "desc": "LNP club uuid"}),
+                        mapper=create_mapper({"id": club.id, "related_type": "club", "database_source": "scrapper_mongodb", "desc": "LNP club uuid"}),
                         scrapper_autocreated=True,
                         mapping=club.name,
                         voivodeship_obj=club_voivo,
@@ -480,9 +407,8 @@ class Command(BaseCommand):
                     )
 
                 def configure_club(obj: Club):
-                    if not obj.mapper:
-                        obj.create_mapper_obj()
-                    MapperEntity.objects.get_or_create(target=obj.mapper, source=NEW_LNP_SOURCE, mapper_id=club.id)
+                    if not obj.mapper or not obj.mapper.get_entities():
+                        obj.mapper = create_mapper({"id": obj.id, "related_type": "club", "database_source": "scrapper_mongodb", "desc": "LNP club uuid"})
                     obj.mapping = club.name
                     if not obj.stadion_address:
                         obj.stadion_address = club_address
@@ -499,13 +425,15 @@ class Command(BaseCommand):
                         club=club_obj,
                         name=team_name,
                         mapper=create_mapper(
-                            NEW_LNP={
+                            {
                                 "id": team_history.obj_id,
+                                "related_type": "team",
+                                "database_source": "scrapper_mongodb",
                                 "desc": "TeamHistory id as ObjectId in mongodb, immutable between seasons"
                             }
                         ),
                         scrapper_autocreated=True,
-                        mapping=f"{team.name}; ",
+                        mapping=f"{team.name};",
                         visible=False,
                         seniority=seniority,
                         gender=gender,
@@ -524,8 +452,10 @@ class Command(BaseCommand):
 
                 if not team_obj.mapper:
                     team_obj.mapper = create_mapper(
-                            NEW_LNP={
+                            {
                                 "id": team_history.obj_id,
+                                "related_type": "team",
+                                "database_source": "scrapper_mongodb",
                                 "desc": "TeamHistory id as ObjectId in mongodb, immutable between seasons"
                             }
                         )
@@ -545,8 +475,7 @@ class Command(BaseCommand):
                         TeamHistory.objects.create(
                             team=team_obj,
                             team_name_raw=team.name,
-                            mapper=create_mapper(NEW_LNP={"id": team.id, "desc": "LNP team uuid"}),
-                            # scrapper_team_uuid=team.id,
+                            mapper=create_mapper({"id": team.id, "related_type": "team history", "database_source": "scrapper_mongodb", "desc": "LNP team uuid"}),
                             league_history=league_history,
                             visible=False,
                         )
@@ -555,25 +484,35 @@ class Command(BaseCommand):
         """
         Merge clubs which are a single instance
         """
+        def define_base_club(qs: List[Club]) -> Club:
+            for club_obj in qs:
+                if club_obj.mapper:
+                    if club_obj.mapper.get_entities():
+                        return club_obj
+            else:
+                return qs[0]
+
         club_list = Club.objects.all()
         for club in club_list:
             same_club_qs = list(Club.objects.filter(name__icontains=club.name))
             if len(same_club_qs) > 1:
-                base_club = same_club_qs[0]
+                base_club = define_base_club(same_club_qs)
                 teams = []
                 editors = []
                 for c in same_club_qs:
                     for team in c.teams.all():
-                        teams.append(team) if team not in teams else False
+                        teams.append(team)
                     for editor in c.editors.all():
-                        editors.append(editor) if editor not in editors else False
-                    editors.append(c.manager) if c.manager else False
-                for team in set(teams):
+                        editors.append(editor)
+                    if c.manager:
+                        editors.append(c.manager)
+                for team in teams:
                     team.club = base_club
                     team.save()
-                for editor in set(editors):
-                    base_club.editors.add(editor)
+                for editor in editors:
+                    if editor not in base_club.editors.all():
+                        base_club.editors.add(editor)
                 base_club.save()
-                same_club_qs.pop(0)
                 for club_to_delete in same_club_qs:
-                    club_to_delete.delete()
+                    if club_to_delete is not base_club:
+                        club_to_delete.delete()
